Add averageReviewByBodyguard

+++ demoProyect/src/main/java/upc/edu/pe/demoproyect/controller/BodyguardController.java
@@ -96,7 +96,7 @@ public class BodyguardController {
     }
 
     @GetMapping("/getReviewBodyguard/{id}")
-    public float getAverageReviewByBodyguardId(@PathVariable int id) {
+    public Integer getAverageReviewByBodyguardId(@PathVariable int id) {
         return bodyguardService.getAverageReviewByBodyguardId(id);
     }
 }
